backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.49
psycopg[binary]==3.2.13
PyJWT==2.8.0
python-multipart==0.0.9
pydantic>=2.0
python-dotenv==1.0.1
supabase==2.7.0
groq==0.9.0
bcrypt==4.1.2
